Fix vsgo bxl builds dropping modifiers before removing toolchain buckconfigs
Summary:
When deleting the toolchain buckconfigs from windows modes, some of the vsgo bxl builds were failing. The reason seems to be that the target isn't picking up the selectified toolchain modifier so it is defaulting to the default toolchain not the selectified one.

This diff passes in the modifiers to the target, so the targets should be setting the windows-arvr-selectified constraint now, which fixes the failing builds

Differential Revision: D85475701

fbshipit-source-id: d70efe5a0b980ebbcf33bdcf9732a0ea4799994f