Allow SDK overrides to support watch targets
Summary:
Currently we can't provide IDE services in Swift files owned by watch targets, because we can't build a compilation database. This is because Apple targets assume an iOS SDK, but watch targets need the watchOS SDK.

Allow Apple targets to override the default SDK, and update our watch targets to use this.

Reviewed By: chatura-atapattu

Differential Revision: D85673943

fbshipit-source-id: 3c3d62aaf911038e031a86d488daa02ea5fae19e