Expose unfiltered argslist of SOCKET manifests
Summary:
As I mentioned [here](https://www.internalfb.com/diff/D85578058?dst_version_fbid=1524731092045549&transaction_fbid=799954839554152), we have an arg on `xplugins_lookup_table` which enables you to pass other `xplugins_lookup_table` manifests we can query for in the graph to prevent plugins from being stripped.

In a world without queries, this is (probably) the unfiltered list of `XPluginsSocketUsageInfo`. Expose it as a subtarget.

Reviewed By: rmaz

Differential Revision: D85635815

fbshipit-source-id: 643b25c46a83f494d5179541c806acd927367aba