Simplify up ShlibInterfacesMode
Summary:
The primary motivation for this change is to remove `ShlibInterfacesMode("enabled")`, because it seems to be un-used and it is confusing because it's not clear which flavor of `ShlibInterfacesMode` that is supposed to enable.

The remaining refactor calculates an "effective_shlib_interfaces_mode" needed to account for the opt outs `ShlibInterfacesMode("defined")` requires. In the next diff, we use this to add a new parameter to `cxx_link_shared_library`, supporting a new `ShlibInterfacesMode`.

Reviewed By: rmaz

Differential Revision: D85153558

fbshipit-source-id: db9f60904ff2a9f37a79b8ab31dee5e58e31e544