Implement linker based shared library interface generation
Summary:
This modifies the linker rules with LinkOption `produce_shared_library_interface` that when set instructs the rules to produce an interface for the shared library it is producing. The resulting interface artifact is passed back to callers as part of the `CxxLinkResult`.

Currently, this only supports Darwin but could be extended for other platforms. With or without distributed thin-lto, the rules create a second action matching the link action already produced, but passing the linker some extra configuration instructing it to produce the shared library interface rather than the link unit itself.

The produced interface is used by rdep link actions as an artifact to link against instead of linking against the actual shared library. This is advantageous for a few reasons:

1. ) It is faster to produce this interface than it is to produce the final link unit. Especially for LTO builds, this can be 10x faster.
2. ) Sometimes we link a link unit twice, the first time just for the sake of producing a shared library to link against, and then again with a dynamically calculated list of exported symbols. In these cases, we can forgo producing a link unit for the first link entirely, and instead use an interface, which saves capacity.

Reviewed By: rmaz

Differential Revision: D85164889

fbshipit-source-id: f0a0af2372059fe129de466589684b75b2173f66