Add activity-alias post processing step before manifest generation
Summary:
At the moment all apps using manifest.py in fbsource already handle how activity-alias are sorted to guarantee they are in the manifest after the activity they target thanks to the post-processing happening in [manifest.py](https://fburl.com/code/fsiow1yu).

In WA this is not the case and we want to guarantee the order is maintained until we are able to update the ManifestMerging plugin to more recent versions. This is a link to a [discussion](https://fburl.com/workplace/byshefjr) regarding this specific manifest merging issue.

Once the latest version of the ManifestMerger will be updated we can clean up this step and the work is tracked in T239793438

Reviewed By: hick209

Differential Revision: D85872474

fbshipit-source-id: ecafa1e77ed1588f3f3933899a4ff520987d423e