Fix support for generated test suites
Summary:
# Context
The `erlang_tests()` macro recieves a list of source files targets in `suites` and generates an `erlang_test` target for each of them. Each target gets the name of the suite.

# Problem
Because the given suite source could be a generated targets, the macro needs to check if the name contains a `:` and do some normalization. Unfortunately, the current normalization is to replace the `:` by ` _`. This is behaviour is surprising to the user as it is not consistent with what happens with plain source files:

- For a target like ":generated_SUITE.erl", the test gets called `"_generated_SUITE.erl"`, which can seem like an error.
- For a target like "foo/bar:generated_SUITE.erl", the target gets called "foo/bar_generaged_SUITE", which is not consistent, as for a file "tests/some_SUITE.erl", we'd generate the test target `some_SUITE`, and not `tests_some_SUITE`.

# This diff
Change the normalization to just take the "base name" of the target

Reviewed By: michalmuskala

Differential Revision: D86523495

fbshipit-source-id: fb3fc7ddc25aa3f26611de3fea525dc5db01f3a2