modular_libraries_use_header_maps disables modulemap preprocessor
Summary:
When `modular_libraries_use_header_maps` is set on the `apple_toolchain`, we only include header symlink tree preprocessor in the `modular_args`. This will exclude it from any cxx compilation of apple_library not using modules.

The Swift rules are extended to always include the modular_args, so that they can see the symlink trees of dependent modules.

Reviewed By: NuriAmari

Differential Revision: D86536249

fbshipit-source-id: 76fcdb478a0c01919bb06638f964cb8292c4a416