Add __kotlin_classes__ output into KotlinExtras
Summary:
### Context
I am trying to minimize the output size generated by KSIC to decrease Buck memory usage associated with directory structure processing.

### This stack
The goal of this stack is to eliminate the duplicated `__classes__` directory in `incremental_state` and instead utilize the `__classes__` output. To achieve this, the Kotlinc output must be isolated by creating a separate folder: `__kotlin_classes__`.

### This diff
Adds the `__kotlin_classes__` folder as the output for `kotlincd_jar`.

Reviewed By: NavidQar

Differential Revision: D86074993

fbshipit-source-id: 264c5f1d1eeeb40eb86e4b9f17426df280d8055b