Use classic Kotlin toolchain with external javac
Summary:
Considering `java_library`, we already force "classic" toolchain compilation (i.e. not javacd) if an explicit `javac` was specified on the target. We do have a handful of targets using this.

This diff does the same for `kotlin_library`. It turns out we have no real world usages of this, but I added a new target to exercise this compilation path.

Basically, this fully drops support for an external javac in the javacd/kotlincd workers - we only allow the jsr199 in-process javac. This enables some further cleanup in that layer.

Reviewed By: NavidQar

Differential Revision: D86803187

fbshipit-source-id: 5d966f3cbd883e2eced481475a7a9b1e14f35b44