Add config for local from od
Summary:
This stack adds a prototype for running Apple tests directly on the user's machine from an OD.

When `buck test` is used in an OD, it normally defaults to remote execution. If you force "local" execution it will fail because it's not possible to create a simulator inside the OD.

This diff adds a new option (`local_from_od`) to the buck config `apple.test_execution` that is for running locally from OD. This option sets up a "local" test run *without* requiring a simulator.

Reviewed By: cpepin

Differential Revision: D86909114

fbshipit-source-id: be857317bc29dc13e866159c7d4e79e0c154cd63