Add an `llvm_stats` subtarget
Summary:
Add an `llvm_stats` subtarget. Only enable it if we are using our `llvm-20` toolchain. The
linker only outputs a stats file iff it runs LTO, which we can't decide at output creation time.
Thus this required a downstream change to our toolchain which is only available on `llvm-20`.

Reviewed By: thevinster

Differential Revision: D87021568

fbshipit-source-id: 5d021d802f95598c8d79846de07486d25d82dfb8