Avoid the list type for accumulating result/stdout
Summary:
# Context
An Erlang test runs in a "test-executor" node, collects all the results in a "results tree", that is saved as an Erlang term to a file. The "test-binary" then reads back that Erlang tree, and converts it into a json file that TPX understands.

The test results tree contains, for each testcase/callback the "stdout" produced in that case and as "details" a readable version of the error term returned by CT in case of failure

# Problem
Both the `details` and the `stdout` are stored as a `string()` in the test-executor, and because there is some concatenation in case of failures in init functions, we keep making copies of these (potentially long lists).

# This diff
Instead use `unicode:chardata()` for this, and remove all the intermediate conversions to string. Because on the test-binary we are already calling `unicode:characters_to_binary()` to convert these fields to json, no other change is needed.

Reviewed By: michalmuskala, TheGeorge

Differential Revision: D87063116

fbshipit-source-id: b4c436d125d8e57b3b76de611802ce6871e13a5a