Avoid processing logs unless there was a crash
Summary:
# Context
When running Erlang tests, we have two nodes, the "test-binary", that starts the "test-executor" node, which actually runs the test and writes results to disk, which the "test-binary" then read back to report, etc.

There are failure modes in which the "test-executor" fails to produce the results file, e.g. because it crashed, or timed out, or bugs. In these cases, we still want to report results, and include in the "stdout" for the test portions of relevant logs: the stdout of the "test-executor" and maybe the logs of the application running the tests inside the "test-executor".

# Problem
We are currently processing these logs even on the happy path, when the "test-executor" succeeded to produce a results file, only to discard it later. This adds latency, but also makes following the logic harder

# This diff
Refactor the logic so that we only do this processing of the logs when we need to.

Reviewed By: michalmuskala

Differential Revision: D86885377

fbshipit-source-id: 6f25c3b30e312922d97b45afd7afc4533f42f143