Use raw file mode when truncating ct_executor stdout file
Summary:
# Context
When running Erlang tests, all stdout for the suite is saved to a file, and its content later reported in some places.

Because this stdout file could be huge, we truncate anything over 5000 characters.

# Problem
When reading the last 5000 chars of the file, we use a iodevice, which is slow. It is much faster to use a raw file descriptor here, and safe since there are no other processes involved.

# This diff
The likely reason why `raw` was not being used is that, against the docs, `file:pread/3` doesn't work on all cases for raw descriptors, but only when the new position is given as an integer 🤷‍♂️

So here we split the `pread()` into a `position()` followed by a `read()`, which will still be faster.

Reviewed By: michalmuskala

Differential Revision: D86867416

fbshipit-source-id: 344cb86ee3429505d2372a4ab2c4b64bfcf8c1e9