Add explicit dependency to cth_hooks app
Summary:
# Context
The cth_hooks application define the cth_tpx hook that we currently use to collect test results.

# Problem
The hook depends on modules in the `common` module, but there is no explicit dependency in the buck2 target. Whlie things will work, it is hard to know quickly what modules one is expected to use from there.

# This diff
Add a explicit dependency

Reviewed By: michalmuskala

Differential Revision: D85760423

fbshipit-source-id: 37d3ed1a8d772bf07883737c7c599e1fa844e246