Back out "Fix constraint value resolution for config_setting targets"
Summary: This backs out D86573972. While the change itself seems to work, we cannot use it easily, since existing tooling depends on config_settings only referencing constraint_values, and not other config_settings. Having this change means that all tooling operating on the buck graph need to handle a case of recursion, which seem more a hassle than what it's worth. Revert this before anyone depends on this.

Reviewed By: JakobDegen

Differential Revision: D87114535

fbshipit-source-id: 7b95e16ca7f9ffd5f864e1044acc837691c3cef8