Support unified constraint for modifiers.conditional
Summary:
Previously we convert the key of conditional modifiers to string by `str(label.raw_target())`.
`label` here is the `ConfiguredProvidersLabel`, `.raw_target()` returns a `TargetLabel` which doesn't have sub target info, which cause the issue for support unified constraint.

So this diff adds support to this.

Since we don't have the direct api to convert  `ConfiguredProvidersLabel` to `ProvidersLabel`, created a util function here.

Reviewed By: scottcao

Differential Revision: D87277101

fbshipit-source-id: 835d743927dcd8d35c857a91695c44e5e73496e2