Pipe Content-Based Flag from Toolchain to Compile Action (second attempt)
Summary:
## Context:

See T239185724. We want to enable content-based output path addressing for Python bytecode.

 ---

## This Diff:

This is another attempt at D84195338.

The previous diff triggered some Buck failures which I suspect were a consequence of a None value getting passed into the content-based path parameter of `declare_output()`.

The field on the Python toolchain can be None:

https://www.internalfb.com/code/fbsource/fbcode/buck2/prelude/python/toolchain.bzl?lines=62

But the corresponding parameter in `declare_output` expects a non-null boolean:

https://www.internalfb.com/code/fbsource/[fcaddbc0ecc2d2a40a42b76bade814e5250600a2]/fbcode/buck2/app/buck2_action_impl/src/context/unsorted.rs?lines=56

So in this revised diff, we explicitly check that the field on the Python toolchain `== True` so that a None value would get passed in as a False to `declare_output()`, avoiding a possible runtime error.

Reviewed By: itamaro

Differential Revision: D87421964

fbshipit-source-id: 81b650a37f69d0de6a12e27b9b74aeeccc06409d