Decouple manifest merger from AOSP and upgrade to `com.android.tools.build:manifest-merger:31.7.3`
Summary:
Meta:
Our manifest merger is outdated (10 years old).
It finally started causing issues that warrant updating it. For example:
- https://fb.workplace.com/groups/2222954841208728/posts/29540156735606227
- https://fb.workplace.com/groups/2222954841208728/posts/29821641977457700

In order for us to be able to upgrade, we had to take following actions:
- Add `package` to the manifest files as it is mandatory in the newer versions of manifest merger: https://fburl.com/diffs/064mzfb9
- Specify an explicit value for 'android:exported': https://fburl.com/diffs/yb69dl96
- Set the targetSdkVersion for library manifests: D87004913

Now we can decouple manifest merger from AOSP and upgrade it, which is what this diff is about.

NOTE: If this diff breaks a build or test, it’s exposing a pre-existing issue hidden by an outdated manifest merger. We should fix forward; do not revert this diff.

Reviewed By: siaojiecai

Differential Revision: D83352021

fbshipit-source-id: b5e4667292293408dd7cd78eea4f4b8f8409d2cd