Provide local and remote executor overrides to tpx
Summary:
This stack moves the `apple.test_execution` buck config to a `tpx` flag. This config is a run time option and should not be invalidating buck state and causing build speed regressions when it changes.

This diff updates the `apple_test` build defs to *always* pass a local and remote command executor override to `tpx`. In a later diff in this stack, `tpx` will use these to setup where a test executes.

Reviewed By: bryanyantx

Differential Revision: D87842176

fbshipit-source-id: 36807820d9fe83f242158b5d984537e95834da91