Replace Content Hash Placeholder With Resolved Hash in Packaging Step
Summary:
## Context:

See T239185724. We want to enable content-based output path addressing for Python bytecode.

 ---

## This Diff:

In the next diff (D84289358), we enable the toggle for content-based path hashing for python binaries/libraries. This caused a lot of test failures, stemming from the fact that the content hash placeholder "`output_artifacts`" was being preserved in the bytecode manifest, causing us to reference non-existent paths when reading from the manifest.

The solution here is to pass in the bytecode artifacts into the packaging step so that we can fetch the resolved content hash and swap out the placeholder for the hash.

 ---

## Impact Analysis:

My analysis here isn't as thorough / complete as I'd like, but in the interest of time, I'll give a rough outline of what the numbers are looking like so far.

**Wins:**

I measured the time it takes for a `dev` build of a large target (`instagram/server/native_python:native_python_ig_server_pre_bolt[standalone]`) after building it once in `opt` mode. The idea is that we would be deduplicating `py_compile` actions between configuration changes, reducing overall build times in these scenarios.

- We consistently see a single additional action build executed, corresponding to the `bytecode_artifacts` write action added in this diff. This write action is relatively cheap as expected (up to about 100ms in CPU time depending on target size.)

- With content based hashing enabled, we're generally seeing an increase in cache hits corresponding the the `py_compile` actions which are now deduplicated. {F1983505506}

 - This seems to correspond to a decrease in wall time for completing the build, though the numbers are very inconsistent in my testing - we will need to observe the rollout to understand the actual impact.

**Guardrails:**

- In all builds, we now see exactly one new action being executed - this corresponds to the `bytecode_artifacts` write action added in this diff. The cost of this write action is fairly small / insignificant as expected. In the worst case (i.e., the largest target I built) this adds about 100ms of CPU time to the build.

- However, we do seem to be increasing overall execution time in a non-trivial way. On the previously mentioned large target, we see execution time consistently increase by about 1%. Some data points:
    - 229:47:25.9s exec time total -> 232:36:38.7s exec time total
    - 228:58:58.8s exec time total -> 230:46:47.2s exec time total
    - 228:59:40.4s exec time total -> 230:51:05.8s exec time total

Reviewed By: christycylee

Differential Revision: D86130268

fbshipit-source-id: ba7557b27e37faa76a0f8df5bc3c916cbec27e83