Stream stdout from test-executor to test-binary
Summary:
# Context
In the last diffs of this stack we've been changing the way we consume stdout when running Erlang tests: we no longer want the "test-executor" (the node running the suite) to try to capture its own stdout and write it to a file) but to "stream" it to the "test-binary" node. This for example, will allow us to show stdout as the test runs, which is useful for debugging, running without TPX, etc.

# This diff
Here we flip the switch and start streaming stdout, following what we did in the previous diff for the test-shell. Main ingredients are:

* When opening the port to the "test-executor" we now use stdin to communicate. We remove the `ReplayIO` option as we don't need to behaviours.
* The ct_runner gen_server now receives stdout lines coming from the port, and saves them to a "raw stdout" file (that later gets processed to remove progress markers, etc).
* On the cth_tpx, which runs in the test-executor, we no longer replace the group-leader with a io process that writes to a file

Reviewed By: TheGeorge

Differential Revision: D87873139

fbshipit-source-id: 382c9d11178534692ee5335e3d881319ab45f2dc