Use collected standard output in testcases reports
Summary:
# Context
At the moment, the logic we use to "capture" stdout in tests is extremely convoluted. The test-runner node hacks up the `user` process to intercept (some, but not all) the produced stdout, and runs some sort of running window of lines.

This has several drawbacks:
  * It is memory heavy (specially since we use strings)
  * The capture logic is not perfect and we tend to omit duplicated lines or miss lines
  * It is not currently possible to "stream" stdout as the test runs

# General solution
We want to move to a state where we instead log progress markers to stdout, and later postprocess the stdout file to find what portion corresponds to each testcase, etc.

# Problem
In D86984397 we have made it so that we can process the raw stdout log and extract the stdout for each testcase. However, we are still reporting the stdout that was collected in the test-runner

# This diff
Here we make it so that we no longer report the stdout collected in test-executor and insted use the one extracted from the logs. This means mostly threading through the map with the collected stdout for each callback.

Later in the stack, we'll remove the logic that collects this on the test-executor

Reviewed By: TheGeorge

Differential Revision: D87178844

fbshipit-source-id: 1ab4048f9c38b85f76193360a6273c3b050d8f28